The One Tower Script
One Tower Script to rule them all, One Tower Script to find them,
One Tower Script to bring them all, and in the darkness bind them

removed all the needless tower scripts the tower to be spawned is
now a paramater passed to the function by the game engine (see
RunScript.Tower (this should be the name of the unit you wish the
ability to spawn) should we need to give each tower custom abilities
in the future (which I expect we will) we can handle this in a case
statment or if it gets to large we can split this to many different
functions in an over arching case statment.
